CAMEL-24560: Address review feedback on token count migration
- Stop narrowing OpenAI embeddings usage headers to int
- Add tests for null/zero Micrometer token counters and Integer factory nulls

+++ components/camel-ai/camel-openai/src/main/java/org/apache/camel/component/openai/OpenAIEmbeddingsProducer.java
@@ -156,8 +156,8 @@ public class OpenAIEmbeddingsProducer extends DefaultAsyncProducer {
         if (response != null) {
             message.setHeader(OpenAIConstants.EMBEDDING_RESPONSE_MODEL, response.model());
             if (response.usage() != null) {
-                message.setHeader(OpenAIConstants.PROMPT_TOKENS, (int) response.usage().promptTokens());
-                message.setHeader(OpenAIConstants.TOTAL_TOKENS, (int) response.usage().totalTokens());
+                message.setHeader(OpenAIConstants.PROMPT_TOKENS, response.usage().promptTokens());
+                message.setHeader(OpenAIConstants.TOTAL_TOKENS, response.usage().totalTokens());
             }
         }
         message.setHeader(OpenAIConstants.EMBEDDING_COUNT, embeddings.size());
